refactor(monocore,monofs): reorganize project structure and improve CLI
- Added NFS server example
- Changed Dir::list() to return Iterator instead of Vec
- Made create_file/dir/symlink methods return specific types
- Added monofs binary and CLI infrastructure
- Added helper functions for CLI testing
- Improved error handling and logging
- Cleaned up dependencies and imports
- Simplified code organization
- Removed monokrun binary and example config files
- Simplified monocore README.md
- Moved filesystem error types to top level in monofs
