+++ .todo.md
@@ -7,18 +7,12 @@
   - [ ] How do we handle "previous" field conflicts?
   - [ ] Handling deletes with tombstones
 
-- [ ] Fix file size issue.
+- [x] Fix file size issue.
 
-  - [ ] We should be able to get the file size from a files metadata.
-  - [ ] This requires making changes to `IpldStore` and `Layout`.
-  - [ ] `IpldStore::get_cid_size(cid: &Cid) -> Result<u64, Error>`?
-  - [ ] `Layout::get_cid_size(cid: &Cid) -> Result<u64, Error>`?
-
-- [ ] Children entities should inherit sync type from their parents when they get added.
-
-- [ ] What if we use LruCache for MonofsServer `fileid_map` and `path_map`?
-
-  - [ ] The issue that comes to mind is that we hand these fileids to clients and they may expect them to be stable.
+  - [x] We should be able to get the file size from a files metadata.
+  - [x] This requires making changes to `IpldStore` and `Layout`.
+  - [x] `IpldStore::get_cid_size(cid: &Cid) -> Result<u64, Error>`?
+  - [x] `Layout::get_cid_size(cid: &Cid) -> Result<u64, Error>`?
 
 ## monoutils-store
