--- monocore.example.toml
@@ -1,24 +1,0 @@
-[[service]]
-name = "sh-counter"
-base = "alpine:latest"
-ram = 256
-cpus = 1
-group = "demo"
-command = "sh"
-args = ["-c", "for i in $(seq 1 20); do echo $i; sleep 2; done"]
-
-[[service]]
-name = "python-counter"
-base = "python:3.11-slim"
-ram = 256
-cpus = 1
-group = "demo"
-command = "python"
-args = [
-    "-c",
-    "import time; count=0; [print(f'Count: {count+1}') or time.sleep(2) or (count:=count+1) for _ in range(20)]",
-]
-
-[[group]]
-name = "demo"
-local_only = true
