+++ monocore/lib/config/defaults.rs
@@ -21,7 +21,3 @@ pub const DEFAULT_SERVER_PORT: u16 = 3456;
 /// The path where all monocore global data is stored.
 pub static DEFAULT_MONOCORE_HOME: LazyLock<PathBuf> =
     LazyLock::new(|| dirs::home_dir().unwrap().join(MONOCORE_HOME_DIR));
-
-/// The path where all monocore project data is stored.
-pub static DEFAULT_MONOCORE_ENV: LazyLock<PathBuf> =
-    LazyLock::new(|| dirs::home_dir().unwrap().join(MONOCORE_HOME_DIR));
