+++ monocore/lib/utils/path.rs
@@ -2,12 +2,9 @@ use std::path::PathBuf;
 
 use monoutils::SupportedPathType;
 
-use crate::{
-    config::{DEFAULT_MONOCORE_ENV, DEFAULT_MONOCORE_HOME},
-    MonocoreError, MonocoreResult,
-};
+use crate::{config::DEFAULT_MONOCORE_HOME, MonocoreError, MonocoreResult};
 
-use super::{MONOCORE_ENV_VAR, MONOCORE_HOME_ENV_VAR};
+use super::MONOCORE_HOME_ENV_VAR;
 
 //--------------------------------------------------------------------------------------------------
 // Constants
@@ -19,8 +16,8 @@ pub const MONOCORE_ENV_DIR: &str = ".menv";
 /// The directory name for monocore's global data
 pub const MONOCORE_HOME_DIR: &str = ".monocore";
 
-/// The directory where project filesystems are stored
-pub const FILESYSTEMS_SUBDIR: &str = "filesystems";
+/// The directory where project root filesystems are stored
+pub const ROOTS_SUBDIR: &str = "roots";
 
 /// The directory where project logs are stored
 pub const LOG_SUBDIR: &str = "log";
@@ -53,15 +50,6 @@ pub fn monocore_home_path() -> PathBuf {
         PathBuf::from(monocore_home)
     } else {
         DEFAULT_MONOCORE_HOME.to_owned()
-    }
-}
-
-/// Returns the path where all monocore project data is stored.
-pub fn monocore_env_path() -> PathBuf {
-    if let Ok(monocore_env) = std::env::var(MONOCORE_ENV_VAR) {
-        PathBuf::from(monocore_env)
-    } else {
-        DEFAULT_MONOCORE_ENV.to_owned()
     }
 }
