+++ monocore/tests/cli/init.rs
@@ -8,6 +8,26 @@ use tempfile::TempDir;
 // Function: Helper
 //--------------------------------------------------------------------------------------------------
 
+/// Get the path to the monocore binary from build directory
+fn get_monocore_bin() -> String {
+    let manifest_dir = std::env::var("CARGO_MANIFEST_DIR").expect("CARGO_MANIFEST_DIR not set");
+    let project_root = Path::new(&manifest_dir)
+        .parent()
+        .expect("Failed to get project root");
+    project_root
+        .join("build")
+        .join("monocore")
+        .to_string_lossy()
+        .into_owned()
+}
+
+/// Create a temporary directory for testing
+fn create_temp_dir() -> (TempDir, std::path::PathBuf) {
+    let temp_dir = TempDir::new().expect("Failed to create temp directory");
+    let temp_path = temp_dir.path().to_path_buf();
+    (temp_dir, temp_path)
+}
+
 /// Helper function to run monocore init command and verify its results
 fn verify_init_results(dir_path: &std::path::Path) {
     // Verify .menv directory was created
@@ -46,20 +66,9 @@ fn verify_init_results(dir_path: &std::path::Path) {
 fn integration_test_init_command_with_path() {
     procspawn::init();
 
-    // Create a temporary directory for the test
-    let temp_dir = TempDir::new().expect("Failed to create temp directory");
-    let temp_path = temp_dir.path().to_path_buf();
-
-    // Get the path to monocore binary from build directory
-    let manifest_dir = std::env::var("CARGO_MANIFEST_DIR").expect("CARGO_MANIFEST_DIR not set");
-    let project_root = Path::new(&manifest_dir)
-        .parent()
-        .expect("Failed to get project root");
-    let monocore_bin = project_root
-        .join("build")
-        .join("monocore")
-        .to_string_lossy()
-        .into_owned();
+    // Create temp directory and get binary path
+    let (_temp_dir, temp_path) = create_temp_dir();
+    let monocore_bin = get_monocore_bin();
 
     // Run the monocore init command with explicit path in a separate process
     let handle = procspawn::spawn((monocore_bin, temp_path.clone()), |(bin, path)| {
@@ -82,141 +91,129 @@ fn integration_test_init_command_with_path() {
     verify_init_results(&temp_path);
 }
 
-// #[test]
-// fn integration_test_init_command_current_dir() {
-//     procspawn::init();
+#[test]
+fn integration_test_init_command_current_dir() {
+    procspawn::init();
 
-//     // Create a temporary directory for the test
-//     let temp_dir = TempDir::new().expect("Failed to create temp directory");
-//     let temp_path = temp_dir.path().to_path_buf();
+    // Create temp directory and get binary path
+    let (_temp_dir, temp_path) = create_temp_dir();
+    let monocore_bin = get_monocore_bin();
 
-//     // Get the path to the monocore binary from environment variable
-//     let monocore_bin =
-//         std::env::var("MONOCORE_BIN").expect("MONOCORE_BIN environment variable not set");
+    // Run the monocore init command without path in a separate process
+    let handle = procspawn::spawn((monocore_bin, temp_path.clone()), |(bin, path)| {
+        let output = Command::new(bin)
+            .arg("init")
+            .current_dir(&path)
+            .output()
+            .expect("Failed to execute monocore init command");
 
-//     // Run the monocore init command without path in a separate process
-//     let handle = procspawn::spawn((monocore_bin, temp_path.clone()), |(bin, path)| {
-//         let output = Command::new(bin)
-//             .arg("init")
-//             .current_dir(&path) // Set working directory instead of changing process working dir
-//             .output()
-//             .expect("Failed to execute monocore init command");
+        assert!(
+            output.status.success(),
+            "monocore init failed: {}",
+            String::from_utf8_lossy(&output.stderr)
+        );
+    });
 
-//         assert!(
-//             output.status.success(),
-//             "monocore init failed: {}",
-//             String::from_utf8_lossy(&output.stderr)
-//         );
-//     });
+    // Wait for the process to complete
+    handle.join().expect("Process failed");
 
-//     // Wait for the process to complete
-//     handle.join().expect("Process failed");
+    verify_init_results(&temp_path);
+}
 
-//     verify_init_results(&temp_path);
-// }
+#[test]
+fn integration_test_init_command_existing_dir() {
+    procspawn::init();
 
-// #[test]
-// fn integration_test_init_command_existing_dir() {
-//     procspawn::init();
+    // Create temp directory and get binary path
+    let (_temp_dir, temp_path) = create_temp_dir();
+    let monocore_bin = get_monocore_bin();
 
-//     // Create a temporary directory for the test
-//     let temp_dir = TempDir::new().expect("Failed to create temp directory");
-//     let temp_path = temp_dir.path().to_path_buf();
+    // Create a file in the directory that shouldn't be affected
+    let test_file = temp_path.join("test.txt");
+    fs::write(&test_file, "test content").unwrap();
 
-//     // Create a file in the directory that shouldn't be affected
-//     let test_file = temp_path.join("test.txt");
-//     fs::write(&test_file, "test content").unwrap();
+    // Run the monocore init command in a separate process
+    let handle = procspawn::spawn(
+        (monocore_bin, temp_path.clone(), test_file.clone()),
+        |(bin, path, test_file)| {
+            let output = Command::new(bin)
+                .arg("init")
+                .arg(&path)
+                .output()
+                .expect("Failed to execute monocore init command");
 
-//     // Get the path to the monocore binary from environment variable
-//     let monocore_bin =
-//         std::env::var("MONOCORE_BIN").expect("MONOCORE_BIN environment variable not set");
+            assert!(
+                output.status.success(),
+                "monocore init failed: {}",
+                String::from_utf8_lossy(&output.stderr)
+            );
 
-//     // Run the monocore init command in a separate process
-//     let handle = procspawn::spawn(
-//         (monocore_bin, temp_path.clone(), test_file.clone()),
-//         |(bin, path, test_file)| {
-//             let output = Command::new(bin)
-//                 .arg("init")
-//                 .arg(&path)
-//                 .output()
-//                 .expect("Failed to execute monocore init command");
+            // Verify the test file still exists and is unchanged
+            assert!(test_file.exists(), "test file should still exist");
+            let contents = fs::read_to_string(&test_file).unwrap();
+            assert_eq!(
+                contents, "test content",
+                "test file contents should be unchanged"
+            );
+        },
+    );
 
-//             assert!(
-//                 output.status.success(),
-//                 "monocore init failed: {}",
-//                 String::from_utf8_lossy(&output.stderr)
-//             );
+    // Wait for the process to complete
+    handle.join().expect("Process failed");
 
-//             // Verify the test file still exists and is unchanged
-//             assert!(test_file.exists(), "test file should still exist");
-//             let contents = fs::read_to_string(&test_file).unwrap();
-//             assert_eq!(
-//                 contents, "test content",
-//                 "test file contents should be unchanged"
-//             );
-//         },
-//     );
+    // Verify monocore files were created
+    assert!(temp_path.join(".menv").exists());
+    assert!(temp_path.join("monocore.yaml").exists());
+}
 
-//     // Wait for the process to complete
-//     handle.join().expect("Process failed");
+#[test]
+fn integration_test_init_command_idempotent() {
+    procspawn::init();
 
-//     // Verify monocore files were created
-//     assert!(temp_path.join(".menv").exists());
-//     assert!(temp_path.join("monocore.yaml").exists());
-// }
+    // Create temp directory and get binary path
+    let (_temp_dir, temp_path) = create_temp_dir();
+    let monocore_bin = get_monocore_bin();
 
-// #[test]
-// fn integration_test_init_command_idempotent() {
-//     procspawn::init();
+    // Run the first init in a separate process
+    let handle = procspawn::spawn((monocore_bin, temp_path.clone()), |(bin, path)| {
+        // First init
+        let output = Command::new(&bin)
+            .arg("init")
+            .arg(&path)
+            .output()
+            .expect("Failed to execute first monocore init command");
 
-//     // Create a temporary directory for the test
-//     let temp_dir = TempDir::new().expect("Failed to create temp directory");
-//     let temp_path = temp_dir.path().to_path_buf();
+        assert!(
+            output.status.success(),
+            "First monocore init failed: {}",
+            String::from_utf8_lossy(&output.stderr)
+        );
 
-//     // Get the path to the monocore binary from environment variable
-//     let monocore_bin =
-//         std::env::var("MONOCORE_BIN").expect("MONOCORE_BIN environment variable not set");
+        // Modify the config file to verify it's not overwritten
+        let config_path = path.join("monocore.yaml");
+        fs::write(&config_path, "modified content").unwrap();
 
-//     // Run the first init in a separate process
-//     let handle = procspawn::spawn((monocore_bin, temp_path.clone()), |(bin, path)| {
-//         // First init
-//         let output = Command::new(&bin)
-//             .arg("init")
-//             .arg(&path)
-//             .output()
-//             .expect("Failed to execute first monocore init command");
+        // Run the second init
+        let output = Command::new(&bin)
+            .arg("init")
+            .arg(&path)
+            .output()
+            .expect("Failed to execute second monocore init command");
 
-//         assert!(
-//             output.status.success(),
-//             "First monocore init failed: {}",
-//             String::from_utf8_lossy(&output.stderr)
-//         );
+        assert!(
+            output.status.success(),
+            "Second monocore init failed: {}",
+            String::from_utf8_lossy(&output.stderr)
+        );
 
-//         // Modify the config file to verify it's not overwritten
-//         let config_path = path.join("monocore.yaml");
-//         fs::write(&config_path, "modified content").unwrap();
+        // Verify the config file wasn't overwritten
+        let contents = fs::read_to_string(&config_path).unwrap();
+        assert_eq!(
+            contents, "modified content",
+            "config file should not be overwritten"
+        );
+    });
 
-//         // Run the second init
-//         let output = Command::new(&bin)
-//             .arg("init")
-//             .arg(&path)
-//             .output()
-//             .expect("Failed to execute second monocore init command");
-
-//         assert!(
-//             output.status.success(),
-//             "Second monocore init failed: {}",
-//             String::from_utf8_lossy(&output.stderr)
-//         );
-
-//         // Verify the config file wasn't overwritten
-//         let contents = fs::read_to_string(&config_path).unwrap();
-//         assert_eq!(
-//             contents, "modified content",
-//             "config file should not be overwritten"
-//         );
-//     });
-
-//     // Wait for the process to complete
-//     handle.join().expect("Process failed");
-// }
+    // Wait for the process to complete
+    handle.join().expect("Process failed");
+}
