+++ monofs/lib/filesystem/cidlink/entity.rs
@@ -1,8 +1,8 @@
 use monoutils_store::IpldStore;
 
-use crate::filesystem::{
-    dir::Dir, entity::Entity, file::File, symcidlink::SymCidLink, CidLink, FsResult, SymPathLink,
-};
+use crate::{filesystem::{
+    dir::Dir, entity::Entity, file::File, symcidlink::SymCidLink, CidLink, SymPathLink,
+}, FsResult};
 
 //--------------------------------------------------------------------------------------------------
 // Types
