+++ monofs/lib/filesystem/mod.rs
@@ -4,7 +4,6 @@ mod cidlink;
 mod dir;
 mod entity;
 mod eq;
-mod error;
 mod file;
 mod kind;
 mod metadata;
@@ -19,7 +18,6 @@ pub use cidlink::*;
 pub use dir::*;
 pub use entity::*;
 pub use eq::*;
-pub use error::*;
 pub use file::*;
 pub use kind::*;
 pub use metadata::*;
